refactor: move enums to separate files for better organization
