+++ app/types/enums/ParticipantEnums.ts
@@ -1,0 +1,15 @@
+export enum ParticipantType {
+  Undefined = "Undefined",
+  PlayerCharacter = "PlayerCharacter",
+  NonPlayerCharacter = "NonPlayerCharacter",
+  Creature = "Creature",
+  Minion = "Minion",
+  Boss = "Boss",
+  Companion = "Companion",
+}
+
+export enum ParticipantStatus {
+  Alive = "Alive",
+  Unconscious = "Unconscious",
+  Dead = "Dead",
+}
